fix(containers): address second round of review feedback on Stats tab
- Move makeSendToFrame into @pkg/window alongside the existing send()
  utility; remove the standalone ipcUtils.ts file
- Clear the poll timer before sending container-stats/stopped so a
  concurrent poll callback cannot fire after the session is torn down
- Add comment explaining why --no-stream is used instead of streaming
- Send stats and process data to the renderer independently so whichever
  arrives first is dispatched without waiting for the other
- Use destructuring in the destroyed / stop handlers to combine the
  existence and sender-identity checks in one guard
- Guard container-stats/stop on sender identity, matching the destroyed
  handler
- Add TODO in ContainerInfo.vue to hide the Stats tab entirely when
  not on the Moby engine (follow-up)
