+++ pkg/rancher-desktop/window/index.ts
@@ -735,3 +735,17 @@ export function centerDialog(window: BrowserWindow, dialog: BrowserWindow, offse
 
   dialog.setPosition(x, y);
 }
+
+/**
+ * Returns a typed helper that sends a renderer IPC event to a specific
+ * WebContents, swallowing errors if the frame has been destroyed.
+ */
+export function makeSendToFrame(sender: Electron.WebContents, logger?: { debug: (...args: any[]) => void }) {
+  return <ch extends keyof IpcRendererEvents>(channel: ch, ...args: Parameters<IpcRendererEvents[ch]>) => {
+    try {
+      sender.send(channel, ...args);
+    } catch (ex) {
+      logger?.debug(`Failed to send ${ channel } to frame:`, ex);
+    }
+  };
+}
